refactor(audit): extract leaf helpers into phi_scan.audit._shared
Move DB/git/CI/timestamp leaf helpers out of the 1437-line audit
monolith into a dedicated _shared module. __init__.py re-exports the
symbols so tests that patch phi_scan.audit._get_current_* and
phi_scan.audit.subprocess.run keep working.

_collect_repository_identity and _collect_committer_identity stay at
the package surface so monkey-patches on the patchable git bindings
apply during identity collection.
